main features ready
adjustments and optional features are not implemented in this commit and will be in the next one. Modules versions issues throw errors in some scripts and functions. I am searching for a fix, cv2 doesn't support latest python version and some modules need a later version to work properly. Due to config loss after a clean install of OS I can't seem to recreate the same, working, environment.